Add files via upload
Task 1:
performing high-level data checks such as:
Creating and interpreting high-level summaries of the data
Finding outliers and removing these 
Checking data formats and correcting 
also deriving extra features such as pack size and brand name from the data and define metrics of interest to enable us to draw insights on who spends on chips and what drives spends for each customer segment.

Task 2:
evaluating the performance of a store trial which was performed in stores 77, 86 and 88.
This can be broken down by:

total sales revenue
total number of customers
average number of transactions per customer
Createing a measure to compare different control stores to each of the trial stores to do this writing a function to reduce having to re-do the analysis for each trial store. Consider using Pearson correlations
Once we have selected the control stores, comparing each trial and control pair during the trial period.  wanting to test if total sales are significantly different in the trial period and if so, checking if the driver of change is more purchasing customers or more purchases per customers etc.